Refactor: move variable_cell.* to module_relax (#1972)
